Fixed generic components bugs, implemented MOS 6502

+++ libemulation/Implementation/Generic/AddressOffset.cpp
@@ -19,7 +19,7 @@ AddressOffset::AddressOffset()
 bool AddressOffset::setValue(string name, string value)
 {
 	if (name == "offset")
-		offset = getInt64(value);
+		offset = getInt(value);
 	else
 		return false;
